refactor: useAuth hook for authentication purpose and redirection in /notes
