+++ notes-web/src/components/core/molecules/Form.tsx
@@ -5,7 +5,9 @@ export const Form = ({ children, onSubmit, title }: any) => {
   return (
     <>
       <h3 className="formTitle">{title}</h3>
-      <form onSubmit={onSubmit}>{children}</form>
+      <form onSubmit={onSubmit}>
+        {children}
+      </form>
     </>
   );
 };
