+++ notes-web/src/components/core/organisms/Header/Header.tsx
@@ -1,47 +1,11 @@
 import './Header.css';
-import React, { useContext, useEffect, useState } from "react";
-import { TokenContext } from "../../../../context";
-import { AuthService } from "../../../../services/auth.service";
-import { LocalStorageUtil } from "../../../../utils/LocalStorageUtil";
+import React from "react";
 import { Navbar } from "../../molecules/Navbar";
 import { LoginForm } from "../Auth/LoginForm";
-import { useHistory } from 'react-router-dom';
+import { useAuth } from '../../../../hooks/useAuth';
 
 export const Header = () => {
-  const history = useHistory();
-  const [ showLogin, setShowLogin ] = useState(false);
-  const { token, setToken } = useContext(TokenContext);
-
-  useEffect(() => {
-    const validateToken = async (authToken: string | null) => {
-      if (authToken) {
-        try {
-          await AuthService.validateToken(authToken);
-          updateTokenAndLogin(authToken);
-        } catch (error) {
-          logout();
-        }
-      }
-    };
-    validateToken(token);
-  }, []);
-
-  const toggleLoginModal = () => {
-    setShowLogin(!showLogin);
-  };
-
-  const updateTokenAndLogin = (authToken: string) => {
-    LocalStorageUtil.set('token', authToken);
-    setToken(authToken);
-    setShowLogin(false);
-    history.push('/notes');
-  };
-
-  const logout = () => {
-    AuthService.logout();
-    setToken('');
-    history.push('');
-  };
+  const { token, showLogin, updateTokenAndLogin, toggleLoginModal, logout } = useAuth();
 
   const navbarElements: { text: string; onClick: () => void; href: string; show: boolean }[] = [
     {
